requirements.txt
streamlit==1.28.0
google-generativeai==0.5.0
python-dotenv==1.0.0
PyPDF2==3.0.1
